// src/shared/interfaces/http-status-codes.ts
// eslint-disable-next-line no-shadow
export enum HttpStatusCode {
  // Success codes
  _200_Ok = 200,
  _201_Created = 201,
  _202_Accepted = 202,
  _203_NonAuthoritativeInformation = 203,
  _204_NoContent = 204,
  _205_ResetContent = 205,
  _206_PartialContent = 206,

  // Redirections
  _300_MultipleChoices = 300,
  _301_MovedPermanently = 301,
  _302_Found = 302,
  _303_SeeOther = 303,
  _304_NotModified = 304,
  _307_TemporaryRedirect = 307,
  _308_PermanentRedirect = 308,

  // Client errors
  _400_BadRequest = 400,
  _401_Unauthorized = 401,
  _402_PaymentRequired = 402,
  _403_Forbidden = 403,
  _404_NotFound = 404,
  _405_MethodNotAllowed = 405,
  _406_NotAcceptable = 406,
  _407_ProxyAuthenticationRequired = 407,
  _408_RequestTimeout = 408,
  _409_Conflict = 409,
  _410_Gone = 410,
  _411_LengthRequired = 411,
  _412_PreconditionFailed = 412,
  _413_PayloadTooLarge = 413,
  _414_RequestURITooLarge = 414,
  _415_UnsupportedMediaType = 415,
  _416_RequestRangeNotSatisfiable = 416,
  _417_ExpectationFailed = 417,
  _418_ImATeapot = 418,
  _421_MisdirectedRequest = 421,
  _422_UnprocessableEntity = 422,
  _423_Locked = 423,
  _424_FailedDependency = 424,
  _426_UpgradeRequired = 426,
  _428_PreconditionRequired = 428,
  _429_TooManyRequests = 429,
  _431_RequestHeaderFieldsTooLarge = 431,
  _451_UnavailableForLegalReasons = 451,

  // Server errors
  _500_InternalServerError = 500,
  _501_NotImplemented = 501,
  _502_BadProvider = 502,
  _503_ServiceUnavailable = 503,
  _504_ProviderTimeout = 504,
  _505_HttpVersionNotSupported = 505,
  _506_VariantAlsoNegotiates = 506,
  _507_InsufficientStorage = 507,
  _508_LoopDetected = 508,
  _510_NotExtended = 510,
  _511_NetworkAuthenticationRequired = 511,
  _599_NetworkConnectionTimeoutError = 599,
}